requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
chromadb==0.4.18
unstructured[pdf]==0.11.8
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0
numpy==1.24.4
python-dotenv==1.0.0
pydantic==2.5.2
sentence-transformers==2.2.2